+++ Experiment/launch_formation/prometheus_formation_control.launch
@@ -1,0 +1,10 @@
+<launch>
+
+    <node pkg="prometheus_mission" type="mocap_formation_control" name="mocap_formation_control" output="screen">
+        <param name="FORMATION_DISTANCE_x" value="2"/>
+        <param name="FORMATION_DISTANCE_y" value="2"/>
+        <param name="Location_source" value="gps"/>
+        <param name="sim" value="$(arg sim)"/>
+    </node>
+
+</launch>
