+++ Experiment/launch_formation/prometheus_formation_setmode.launch
@@ -1,0 +1,10 @@
+<launch>
+
+    <node pkg="prometheus_mission" type="mocap_formation_setmode" name="mocap_formation_setmode" output="screen">
+        <param name="OFFBOARD_intervals" value="0"/>
+        <param name="LAND_intervals" value="0"/>
+        <param name="Flight_controller" value="px4"/>
+        <param name="Takeoff_height" value="1"/>
+    </node>
+
+</launch>
